Add find pid program
This program is to find or check a process in /proc,
when given a process name for all the running the process,
if we don't know the process pid that we execute or executed.
It trys to read and check each of a list of pid with the /proc/pid/stat.

Signed-off-by: Myoungje Kim <mjei78@gmail.com>